Now passes all 3 QueenMoveTests()
This was even faster than the last one, and I hope this doesn't trigger the suspiciously frequent commits flag. All it took was adding to a queen all the moves that apply to a rook and bishop. I'm actually really happy with the concept of this change because it saved a lot of work.

If and TA has questions or lingering suspicion I'm happy to talk to them.  :)