Passes all tests except extra credit and FullGame.
The error in FullGame is that the lists of current pieces aren't updated when a regular move is performed, so they quickly become inaccurate. I'm about to fix that, but I haven't committed in a while so I figured I'd do that to save all of my other progress in case this messes something up because everything else is working great.